Pull out cache into separate file
